fixed weird memory bug in the default move constructor of CellBase
Summary: turns out that the default move container in combination with clang segfaults under certain (unclear) cicumstances, because the move constructor of the optional appears to be busted in gcc 7.2. Copying it (K) instead of moving it fixes the issue, and since it is a reference, the cost is practically nil.

Test Plan: CI passes

Reviewers: afalsafi

Reviewed By: afalsafi

Differential Revision: https://c4science.ch/D188