Fixed documentation build on readthedocs.io
Summary: At some point, the readthedocs builds failed when interpreting CMakeLists.txt files as c++ code. They are now excluded from doxygen (as they probably should have been in the first place)

Test Plan: None of the tests or builds are affected by this. The only meaningful test is to see that the documentation for "latest" now builds again on muspectre.readthedocs.io)

Reviewers: afalsafi

Reviewed By: afalsafi

Differential Revision: https://c4science.ch/D189