ENH: Native stress can be stored
The `compute_stresses` and `compute_stresses_tangent` functions of materials now
take an additional argument of enum type `StoreNativeStress` (which can be `yes`
or `no`). If yes, the native stress (e.g., Kirchhoff stress for many plasticity
laws, or Material/Piola-Kirchhoff-2 stress for elastic laws) is stored in a
local field called "native stress" to be used for visualisation or specialty
solvers such as the discrete plasticity law by @rleute.